requirements.txt
asgiref==3.8.1
Django==5.1.7
django-modeltranslation==0.19.13
django-unfold==0.52.0
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.1
